Add minimal reproducer
Add a markdown content with inline code and show the excerpt
in index page.

We can see that the text in inline code is pruned in excerpt. It
seems that `gatsby-remark-prismjs` will transform the `inlineCode`
node as `html` node with `<code>{...}</code>` so `gatsby-
transformer-remark` will ignore it when extracting excerpt.